Move unc_url function to url.cpp
UNC conversions are also needed for URLHandler hence it is more natural
to have this function placed with other url functions.
