+++ test/CMakeLists.txt
@@ -18,7 +18,6 @@ set(TEST_SRCS
     test_graph.cpp
     test_pinning.cpp
     test_virtual_packages.cpp
-    test_fetch.cpp
 )
 
 add_executable(test_mamba ${TEST_SRCS})
